Add basic player movement;
