+++ RgbJourney/RgbJourney/Game1.cs
@@ -91,28 +91,28 @@ namespace RgbJourney
             SpriteBatch = new SpriteBatch(GraphicsDevice);
             //_resourceManager = new ResourceManager(GraphicsDevice, Content);
 
-            ResetGameState();
+            //ResetGameState();
             
 
             // TODO: use this.Content to load your game content here
         }
 
-        private void ResetGameState()
-        {
-            _fieldManager = new FieldManager(_cellSize, _cellSpacing, SpriteBatch, _resourceManager, _fieldSize);
+        //private void ResetGameState()
+        //{
+        //    _fieldManager = new FieldManager(_cellSize, _cellSpacing, SpriteBatch, _resourceManager, _fieldSize);
 
-            //Alpha-2 refactor
-            _fieldGenerator = new FieldGenerator(_cellSize, _cellSpacing, _fieldSize, _random);
-            _field = _fieldGenerator.GenerateField();
-            _gameStepManager = new GameStepManager(_random);
-            //----
+        //    //Alpha-2 refactor
+        //    _fieldGenerator = new FieldGenerator(_cellSize, _cellSpacing, _fieldSize, _random);
+        //    _field = _fieldGenerator.GenerateField();
+        //    _gameStepManager = new GameStepManager(_random);
+        //    //----
 
-            _player = new Player(_cellSize, _cellSpacing, _fieldSize, SpriteBatch, _resourceManager);
-            _fieldManager.OldPlayerPosition = new Position(_player.Position);
-            _uiManager = new UIManager(_cellSize, _cellSpacing,
-                GraphicsDevice.Viewport.Width,
-                GraphicsDevice.Viewport.Height, _fieldSize, SpriteBatch, _resourceManager);
-        }
+        //    _player = new Player(_cellSize, _cellSpacing, _fieldSize, SpriteBatch, _resourceManager);
+        //    _fieldManager.OldPlayerPosition = new Position(_player.Position);
+        //    _uiManager = new UIManager(_cellSize, _cellSpacing,
+        //        GraphicsDevice.Viewport.Width,
+        //        GraphicsDevice.Viewport.Height, _fieldSize, SpriteBatch, _resourceManager);
+        //}
 
         protected override void Update(GameTime gameTime)
         {
@@ -151,7 +151,7 @@ namespace RgbJourney
                 || keyboardNewState.IsKeyDown(Keys.D1) && !_keyboardOldState.IsKeyDown(Keys.D1))
             {
                 _fieldSize = 15;
-                ResetGameState();
+                //ResetGameState();
                 _gamePhase = GamePhase.Game;
                 _gameStep = GameStep.First;
                 _gameStartSeconds = gameTime.TotalGameTime.TotalSeconds;
@@ -160,7 +160,7 @@ namespace RgbJourney
                 || keyboardNewState.IsKeyDown(Keys.D2) && !_keyboardOldState.IsKeyDown(Keys.D2))
             {
                 _fieldSize = 9;
-                ResetGameState();
+                //ResetGameState();
                 _gamePhase = GamePhase.Game;
                 _gameStep = GameStep.First;
                 _gameStartSeconds = gameTime.TotalGameTime.TotalSeconds;
@@ -239,7 +239,7 @@ namespace RgbJourney
         private void DrawGameScreen(GameTime gameTime)
         {
             _fieldManager.DrawField(_field);
-            _player.Draw();
+            //_player.Draw();
             _uiManager.Draw();
             if (_gameStep != GameStep.First)
             {
