+++ RgbJourney/RgbJourney/GameScreens/GamePlayScreen.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework.Input;
 using RgbJourney.Controls;
 using RgbJourney.Models;
 using System;
@@ -19,8 +20,10 @@ namespace RgbJourney.GameScreens
         FieldManager _fieldManager;
         ResourceManager _resourceManager;
         Field _field;
+        Player _player;
 
         private PictureBox backgroundImage;
+        private Texture2D playerTexture;
 
         public GamePlayScreen(Game game, GameStateManager stateManager) : base(game, stateManager)
         {
@@ -37,6 +40,9 @@ namespace RgbJourney.GameScreens
 
             ControlManager.Add(backgroundImage);
 
+            playerTexture = content.Load<Texture2D>("Player");
+
+            _player = new Player(_cellSize, _cellSpacing, _fieldSize, playerTexture);
             _resourceManager = new ResourceManager();
             _fieldGenerator = new FieldGenerator(_cellSize, _cellSpacing, _fieldSize, new Random());
             _field = _fieldGenerator.GenerateField();
@@ -55,12 +61,22 @@ namespace RgbJourney.GameScreens
             base.Draw(gameTime);
             ControlManager.Draw(GameRef.SpriteBatch);
             _fieldManager.DrawField(_field);
+            _player.Draw(GameRef.SpriteBatch);
             GameRef.SpriteBatch.End();
         }
 
         public override void Update(GameTime gameTime)
         {
             base.Update(gameTime);
+
+            if (InputHandler.KeyPressed(Keys.Up))
+                _player.MoveUp();
+            if (InputHandler.KeyPressed(Keys.Down))
+                _player.MoveDown();
+            if (InputHandler.KeyPressed(Keys.Left))
+                _player.MoveLeft();
+            if (InputHandler.KeyPressed(Keys.Right))
+                _player.MoveRight();
         }
     }
 }
